chronoreach-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
aiosqlite==0.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
sse-starlette==2.0.0
pandas==2.2.1
openpyxl==3.1.2
twilio==9.0.2
openai==1.14.2
python-multipart==0.0.9
google-generativeai>=0.8.0
groq>=0.11.0
tavily-python>=0.5.0
tenacity>=8.2.0
aiofiles>=23.2.0
requests>=2.31.0
